requirements.txt
Flask~=2.2.2
python-magic~=0.4.27